update directions almost finished

+++ src/app/component/cards/cards.component.ts
@@ -14,6 +14,7 @@ export class CardsComponent implements OnInit {
   isVisibleMiddle = false;
   @ViewChild(AddArtworkCardComponent) addArtWorkCardComponent: AddArtworkCardComponent;
 
+  // tslint:disable-next-line:no-shadowed-variable
   constructor(private ImageStorageService: ImageStorageService, private ArtWorkService: ArtWorkService) {
   }
 
@@ -22,7 +23,7 @@ export class CardsComponent implements OnInit {
   ArtWorkImageLink = 'https://images.bluethumb.com.au/uploads/listing/201464/debra-ryan-chakra-feather-modern-inspirational-artwork-bluethumb-01c5.jpg?w=766&h=766&fit=crop&auto=compress&cs=tinysrgb&q=70&s=9e0a3ef9469be01f1bc75e1e615d5caa';
 
   ngOnInit(): void {
-    let t = new Date(1970, 0, 1);
+    const t = new Date(1970, 0, 1);
     // @ts-ignore
     t.setSeconds(this.artWork.Date.seconds);
     this.artWork.Date = t;
@@ -47,13 +48,27 @@ export class CardsComponent implements OnInit {
     const Description = this.addArtWorkCardComponent.NameInput.value;
     const Title = this.addArtWorkCardComponent.DescriptionInput.value;
     const Author = this.addArtWorkCardComponent.AuthorInput.value;
+
     const image = this.addArtWorkCardComponent.UploaderInput.image;
-    const Urimage = image.name;
+    console.log(image)
+    const Urimage = (image?.name)?image.name:this.artWork.Urimage;
+    console.log(Urimage)
     const artWork = {Author, Date: new Date(this.addArtWorkCardComponent.DateInput.value), Description, Title, Urimage};
-    this.ImageStorageService.uploadImage('ArtImages', image.name, image).then(res =>
-      this.ArtWorkService.updateArtWorks(this.artWork.id, artWork)
-    );
-    this.ImageStorageService.deletePreviousImage('ArtImages', this.artWork.Urimage).then();
+    console.log(artWork);
+
+    if (image){
+      console.log("oui")
+      this.ImageStorageService.uploadImage('ArtImages', image.name, image).then(res =>
+          this.ArtWorkService.updateArtWorks(this.artWork.id, artWork)
+        );
+      this.ImageStorageService.deletePreviousImage('ArtImages', this.artWork.Urimage).then();
+    }
+    else{
+      console.log("non")
+      this.ArtWorkService.updateArtWorks(this.artWork.id, artWork);
+    }
+
+
     this.isVisibleMiddle = false;
   }
